+++ frontend/src/main/scala/pl/edu/agh/sius/rpc/RPCService.scala
@@ -3,7 +3,7 @@ package pl.edu.agh.sius.rpc
 import pl.edu.agh.sius.services.PiService
 
 class RPCService extends MainClientRPC {
-  override def update(totalPoints: Int, pointsInCircle: Int): Unit =
+  override def update(totalPoints: Double, pointsInCircle: Double): Unit =
     PiService.update(totalPoints, pointsInCircle)
 }
